requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python>=4.6.0
Pillow>=9.0.0
albumentations>=1.3.0
efficientnet-pytorch>=0.7.0
timm>=0.9.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.25.0
tensorboard>=2.10.0
tqdm>=4.60.0
wandb>=0.15.0
